separated primary doc functions
