+++ parser/parser_13f_primary_doc_xml.py
@@ -2,26 +2,24 @@ import requests
 from xml.etree import ElementTree
 
 
-def primary_doc_parser(primary_doc_xml):
-    url = primary_doc_xml
-    getter = requests.get(url)
+def grab_primary_doc_root(primary_doc_xml):
+    getter = requests.get(primary_doc_xml)
     text = getter.text
     root = ElementTree.XML(text)
+    return root
 
-    # grab CIK & company name
-    # TODO; dynamically create namespaces
+
+def grab_primary_doc_cik(root):
     namespaces = {'original': 'http://www.sec.gov/edgar/thirteenffiler',
                   'ns1': 'http://www.sec.gov/edgar/common'}
-
     for cik in root.findall('original:headerData/original:filerInfo/original:filer/original:credentials/original:cik',
                             namespaces):
-        # print(cik.text)
-        yield cik.text
+        return cik.text
 
+
+def grab_primary_doc_company_name(root):
+    namespaces = {'original': 'http://www.sec.gov/edgar/thirteenffiler',
+                  'ns1': 'http://www.sec.gov/edgar/common'}
     for company_name in root.findall('original:formData/original:coverPage/original:filingManager/original:name',
                                      namespaces):
-       # print(company_name.text)
-       # return company_name.text
-       pass
-
-    # TODO; return cik and company name
+        return company_name.text
